+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- `lending-iter` UNSTABLE feature that introduce lending iterator using GATs
+
 ## [1.0.0] - 2022-11-19
 
 ### Changed